feat:landing page tranfer to components base landing page and add scroll-based section trigger animations
