+++ frontend/package.json
@@ -20,6 +20,7 @@
     "react-helmet": "^6.1.0",
     "react-hot-toast": "^2.5.2",
     "react-icons": "^5.5.0",
+    "react-intersection-observer": "^9.16.0",
     "react-router-dom": "^7.6.2",
     "styled-components": "^6.1.19",
     "zustand": "^5.0.6"
