+++ frontend/src/components/ui/AnimatedSection.jsx
@@ -1,0 +1,21 @@
+// components/ui/AnimatedSection.jsx
+import React from 'react';
+import { motion } from 'framer-motion';
+import { useInView } from 'react-intersection-observer';
+
+const AnimatedSection = ({ children, delay = 0.2 }) => {
+  const { ref, inView } = useInView({ triggerOnce: true, threshold: 0.2 });
+
+  return (
+    <motion.div
+      ref={ref}
+      initial={{ opacity: 0, y: 50 }}
+      animate={inView ? { opacity: 1, y: 0 } : {}}
+      transition={{ duration: 0.6, delay }}
+    >
+      {children}
+    </motion.div>
+  );
+};
+
+export default AnimatedSection;
